Moved models.py into app. Renamed common/models to pseudo_models:
- Refactored imports for app init to configure these changes
